Organized cloud-sql-python-connector into namespace module (#4)
* Created module folder for cloud-sql-python-connector.

* Moved folders to namespace style, modified setup.py and noxfile.py to adjust for that

* Removed Python2.7 and Python3.5 runtimes from testing since they are not supported.

* Made a connector folder.
