+++ .travis.yml
@@ -6,8 +6,6 @@ python:
 before_install:
     - sudo add-apt-repository ppa:deadsnakes/ppa -y
     - sudo apt-get update
-    - sudo apt-get install python2.7
-    - sudo apt-get install python3.5
     - sudo apt-get install python3.6
 
 install:
